# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Handdraw API",
    "version": "1.0.0",
    "description": "Handdraw API contract. Identity, workspace onboarding and project/board routes are wired behind explicit configuration. Board deletion additionally requires the cleanup worker. Team membership, invitations and board Viewer grants require the membership feature flag. Invitation email delivery remains pending integration. Bearer identity is verified before profile resolution. Unknown fields/trailing JSON are rejected. Future domain endpoints are added with their implementation contracts."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Local development; deployment must use HTTPS."
    }
  ],
  "security": [
    {
      "supabaseBearer": []
    }
  ],
...
  "paths": {
...
    "/v1/me/shared-boards": {
      "get": {
        "operationId": "listSharedBoards",
        "summary": "List shared boards",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Boards"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available with APP_MEMBERSHIP_ENABLED, board/workspace/identity enabled and schema version 9. Available when the corresponding runtime feature is enabled. Cursor ordering: resource ID ascending."
      }
    },
...
    "/v1/workspaces/{workspace_id}/projects": {
      "get": {
        "operationId": "listProjects",
        "summary": "List projects",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspace_id"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Projects"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available when the corresponding runtime feature is enabled."
      },
      "post": {
        "operationId": "createProject",
        "summary": "Create project",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspace_id"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "pattern": "^\"[1-9][0-9]*\"$"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          },
          "202": {
            "description": "Operation already processing; result contains a safe operation reference. No replay of cached secrets.",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingOperationResponse"
                }
              }
            }
          }
        },
        "x-max-request-bytes": 65536,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          }
        },
        "tags": [
          "Projects"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available when the corresponding runtime feature is enabled."
      }
    },
    "/v1/projects/{project_id}": {
      "patch": {
        "operationId": "renameProject",
        "summary": "Rename project",
        "parameters": [
          {
            "$ref": "#/components/parameters/project_id"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "pattern": "^\"[1-9][0-9]*\"$"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          },
          "412": {
            "$ref": "#/components/responses/412"
          },
          "428": {
            "$ref": "#/components/responses/428"
          }
        },
        "x-max-request-bytes": 65536,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rename"
              }
            }
          }
        },
        "tags": [
          "Projects"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available when the corresponding runtime feature is enabled."
      },
      "delete": {
        "operationId": "deleteEmptyProject",
        "summary": "Delete empty project",
        "parameters": [
          {
            "$ref": "#/components/parameters/project_id"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "responses": {
          "204": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          },
          "412": {
            "$ref": "#/components/responses/412"
          },
          "428": {
            "$ref": "#/components/responses/428"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Projects"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available when the corresponding runtime feature is enabled. Reject nonempty projects; repeat delete with parent access has no additional effect."
      }
    },
...
    "/v1/boards/{board_id}/document": {
      "get": {
        "operationId": "getBoardDocument",
        "summary": "Get board document",
        "parameters": [
          {
            "$ref": "#/components/parameters/board_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              },
              "X-Document-Schema-Version": {
                "schema": {
                  "type": "integer",
                  "enum": [
                    1
                  ]
                }
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Boards"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available when the corresponding runtime feature is enabled. Read committed schema-1 Yjs update bytes; no offline or mutable REST document writes."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "supabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Supabase JWT"
      }
    },
    "schemas": {
...
      "BoardListResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "meta": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "request_id": {
                "type": "string",
                "minLength": 1
              },
              "pagination": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "next_cursor": {
                    "type": "string",
                    "maxLength": 4096,
                    "nullable": true
                  }
                },
                "required": [
                  "next_cursor"
                ]
              }
            },
            "required": [
              "request_id",
              "pagination"
            ]
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          }
        },
        "required": [
          "success",
          "meta",
          "result"
        ]
      },
...
    },
    "parameters": {
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        },
        "description": "Opaque authenticated cursor bound to actor, workspace, resource, filter and the operation-specific ordering described below; mismatch is invalid_cursor."
      },
...
      "workspace_id": {
        "name": "workspace_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/WorkspaceID"
        }
      },
      "project_id": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/ProjectID"
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "invalid_request / invalid_cursor / invalid_resource_id",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "401": {
        "description": "unauthenticated",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "403": {
        "description": "permission_denied / entitlement_read_only",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}